update DNATaxo and PST

+++ code/module/pstRunner.py
@@ -1,7 +1,10 @@
-import subprocess
 import os
-import shutil
 import h5py
+import math
+import time
+import torch
+import shutil
+import subprocess
 
 
 from entity.sample import Sample
@@ -25,24 +28,40 @@ class PSTRunner(DNALMRunner):
         for sample in samples:
             if (len(sample.proteins) > 1):
                 samplesToRun.append(sample)
-        tmpFolder = f"{config.cacheFolder}/PST"
-        inputFile = f"{tmpFolder}/input.fasta"
-        outputFile = f"{tmpFolder}/embeddings.h5"
-        os.makedirs(tmpFolder, exist_ok=True)
-        IOUtils.writeSampleProteinFasta(samplesToRun, inputFile, withHead=True)
+        GPUs = torch.cuda.device_count()
+        samplesPerGPU = math.ceil(len(samplesToRun) / GPUs)
 
-        cmd = f"conda run -n pst --no-capture-output python main.py {inputFile} {tmpFolder} {self.model}"
-        cwd = "/Software/protein_set_transformer"
-        subprocess.run(cmd, cwd=cwd, shell=True)
+        processes:list[subprocess.Popen] = []
 
+        for i in range(GPUs):
+            tmpFolder = f"{config.cacheFolder}/PST_{i}"
+            inputFile = f"{tmpFolder}/input.fasta"
+            os.makedirs(tmpFolder, exist_ok=True)
+            IOUtils.writeSampleProteinFasta(samplesToRun[i * samplesPerGPU : (i+1) * samplesPerGPU], inputFile, withHead=True)
 
-        file = h5py.File(outputFile, "r")
+            env = os.environ.copy()
+            env['CUDA_VISIBLE_DEVICES'] = str(i)
+            cmd = f"conda run -n pst --no-capture-output python main.py {inputFile} {tmpFolder} {self.model}"
+            cwd = "/Software/protein_set_transformer"
+            processes.append((i, subprocess.Popen(cmd, cwd=cwd, shell=True, env=env)))
+        # subprocess.run(cmd, cwd=cwd, shell=True)
 
-        emb = file["genome"][:]  # Load into numpy array
-        for idx, s in enumerate(samplesToRun):
-            embeddings[s.id] = emb[idx, :]
+        while processes:
+            for i, p in processes[:]:
+                if p.poll() is not None:
+                    tmpFolder = f"{config.cacheFolder}/PST_{i}"
+                    outputFile = f"{tmpFolder}/embeddings.h5"
+                    file = h5py.File(outputFile, "r")
 
-        shutil.rmtree(tmpFolder)
+                    emb = file["genome"][:]  # Load into numpy array
+                    for idx, s in enumerate(samplesToRun[i * samplesPerGPU : (i+1) * samplesPerGPU]):
+                        embeddings[s.id] = emb[idx, :]
+
+                    shutil.rmtree(tmpFolder)
+
+                    processes.remove((i, p))
+            time.sleep(1)
+
         return embeddings
     
     def clean(self):
